model-deployment/requirements.txt
pandas
scikit-learn
tensorflow
confluent-kafka
avro-python3
requests
